Fixes a bug where local SQL data was not created after successful server response
This bug sometimes appeared on `collection.fetch` with  `opts.deleteAllOnFetch`.

The server answered correctly but it didn't update the local database. I think this only happens when the queryList is used. If it inserts directly, all is fine. 

When I was debugging this I found that the iteration function gets an Array of arguments passed (the data itself) where i should be passed. That happens on the first iteration call, where i should be set to null but _.defer doesn't pass nothing (undefined) as second parameter but these arguments.